+++ src/main/java/eu/iv4xr/framework/extensions/occ/UserCharacterization.java
@@ -20,9 +20,11 @@ public abstract class UserCharacterization {
     /**
      * Appraise the desirability of an event towards a goal. Positive is desirable,
      * negative is the opposite. 0 means the event is considered as not relevant
-     * towards the goal.
+     * towards the goal. 1 id the event is very desirable.
+     * 
+     * <p>Desirability should be a value in [-1..1]
      */
-    public abstract int desirabilityAppraisalRule(Goals_Status goals_status, String eventName, String goalName);
+    public abstract double desirabilityAppraisalRule(Goals_Status goals_status, String eventName, String goalName);
 
     /*
      * Commenting this out. These are not really goal-oriented appraisal... public
@@ -38,12 +40,13 @@ public abstract class UserCharacterization {
 
     /**
      * This rule specifies for each emotion type, what the corresponding
-     * decay-factor. This should be an int >= 1. Emotion with decay-factor 2 will
-     * decay, roughly, twice as fast.
+     * decay-factor. This should be a number >= 1. The bigger the factor, the faster the emotion
+     * decay over time. The definition of the actual decay function used can be found here:
+     * {@link EmotionRelatedFunctions#decayedIntesity(UserCharacterization, eu.iv4xr.framework.extensions.occ.Emotion.EmotionType, double, long, long)}.
      */
-    public abstract int emotionIntensityDecayRule(Emotion.EmotionType etype);
+    public abstract double emotionIntensityDecayRule(Emotion.EmotionType etype);
 
-    public abstract int intensityThresholdRule(Emotion.EmotionType etyp);
+    public abstract double intensityThresholdRule(Emotion.EmotionType etyp);
 
     /**
      * Representing a rule (actually, a combined set of rules) describing how the
@@ -52,8 +55,8 @@ public abstract class UserCharacterization {
      * likelihood of the consequence goal. Null is returned if the causing-goal is
      * considered as irrelevant for the consequent-goal.
      * 
-     * The default is that this rule is just empty.
+     * <p>The default is that this rule is just empty.
      */
-    public Function<BeliefBase, Function<String, Function<String, Integer>>> goalTowardsGoalRule = bbs -> cause -> consequent -> null;
+    public Function<BeliefBase, Function<String, Function<String, Double>>> goalTowardsGoalRule = bbs -> cause -> consequent -> null;
 
 }
